style(inventory): :art: productID changed by product_id
