+++ src/inventory/application/commands/create/types/dto.py
@@ -4,8 +4,8 @@ from inventory.application.models.inventory import UserRole
 class CreateInventoryDto:
     def __init__(
         self,
-        productId: int,
+        product_id: int,
         stock: int,
     ):
-        self.productId = productId
+        self.productId = product_id
         self.stock = stock
