+++ src/inventory/infraestructure/models/postgres/sqlalchemy/inventory_model.py
@@ -1,0 +1,10 @@
+from sqlalchemy import Column, ForeignKey, int, String
+from common.infrastructure.database.database import Base
+
+
+class ProductModel(Base):
+    __tablename__ = "inventory"
+
+    id = Column(String, primary_key=True, index=True)
+    product_id = Column(ForeignKey('products.id'))
+    stock = Column(int)
